+++ wow_world_base/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+* `Aura` types for Wrath and TBC.
 
 ## [0.2.0] - 2024-02-21